testes apis

+++ apis/betterplacebcd/src/main/java/betterplace/betterplacebcd/controle/CampanhaController.java
@@ -18,8 +18,10 @@ public class CampanhaController{
     private CampanhaRepository repository;
 
     @PostMapping
-    public ResponseEntity postCampanha(@RequestBody @Valid Campanha novaCampanha) {
+    public  ResponseEntity postCampanha(@RequestBody @Valid Campanha novaCampanha) {
         novaCampanha.setDataCriacao(LocalDateTime.now());
+        if (novaCampanha.getNomeCampanha() == null || novaCampanha.getDescCampanha() == null || novaCampanha.getValorNecessario() == null)
+            return ResponseEntity.status(400).build();
         repository.save(novaCampanha);
         return ResponseEntity.status(201).build();
     }
@@ -45,6 +47,9 @@ public class CampanhaController{
 
     @DeleteMapping("/{cod}")
     public ResponseEntity apagarVakinha(@PathVariable Integer cod){
+        if(!repository.findByCodigo(cod)){
+            return ResponseEntity.status(404).build();
+        }
         repository.delete(repository.getById(cod));
         return ResponseEntity.status(201).build();
     }
